requirements.txt
pandas==2.2.2
numpy==1.26.4
Faker==30.8.2
python-dotenv==1.0.1
psycopg[binary]==3.2.3
SQLAlchemy==2.0.36
tqdm==4.66.6
pyarrow==17.0.0; python_version < "3.13"
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
statsmodels==0.14.4
openpyxl==3.1.5
fastapi==0.115.6
uvicorn[standard]==0.32.1
pytest==8.3.4
httpx==0.27.2
